+++ src/main/java/spoticks/ticket_reservation/global/auth/dto/AuthRequest.java
@@ -1,4 +1,4 @@
-package spoticks.ticket_reservation.global.auth;
+package spoticks.ticket_reservation.global.auth.dto;
 
 import lombok.Getter;
