+++ src/main/java/spoticks/ticket_reservation/global/auth/service/AuthorizationUtil.java
@@ -1,8 +1,9 @@
-package spoticks.ticket_reservation.global.auth;
+package spoticks.ticket_reservation.global.auth.service;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
+import spoticks.ticket_reservation.global.auth.entity.CustomUserDetails;
 
 @Component
 public class AuthorizationUtil {
