+++ src/main/java/spoticks/ticket_reservation/global/auth/service/CustomUserDetailsService.java
@@ -1,10 +1,11 @@
-package spoticks.ticket_reservation.global.auth;
+package spoticks.ticket_reservation.global.auth.service;
 
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import spoticks.ticket_reservation.domain.member.entity.Member;
 import spoticks.ticket_reservation.domain.member.repository.MemberRepository;
+import spoticks.ticket_reservation.global.auth.entity.CustomUserDetails;
 
 @Service
 public class CustomUserDetailsService implements UserDetailsService {
